[stats] Move jobstats support code to its own module.
